requirements.txt
streamlit
requests
pandas
python-dotenv
google-generativeai>=0.3.0